+++ src/main/java/com/yanny/ytech/configuration/block_entity/BronzeAnvilBlockEntity.java
@@ -3,12 +3,12 @@ package com.yanny.ytech.configuration.block_entity;
 import com.yanny.ytech.configuration.recipe.HammeringRecipe;
 import com.yanny.ytech.registration.YTechBlockEntityTypes;
 import com.yanny.ytech.registration.YTechRecipeTypes;
+import com.yanny.ytech.registration.YTechSoundEvents;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.protocol.Packet;
 import net.minecraft.network.protocol.game.ClientGamePacketListener;
 import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
-import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.Container;
 import net.minecraft.world.Containers;
@@ -56,7 +56,6 @@ public class BronzeAnvilBlockEntity extends BlockEntity {
                 Function<HammeringRecipe, Float> getStep = (recipe) -> 1F;
                 BiConsumer<Container, HammeringRecipe> onFinish = (container, recipe) -> {
                     Containers.dropItemStack(level, pos.getX(), pos.getY(), pos.getZ(), recipe.assemble(container, level.registryAccess()));
-                    level.playSound(null, pos, SoundEvents.ANVIL_USE, SoundSource.BLOCKS, 1.0f, 0.8f + level.random.nextFloat() * 0.4f);
                 };
 
                 if (!progressHandler.tick(level, canProcess, getStep, onFinish)) {
@@ -64,7 +63,7 @@ public class BronzeAnvilBlockEntity extends BlockEntity {
                     progressHandler.clear();
                 } else {
                     player.getItemInHand(hand).hurtAndBreak(1, player, (e) -> e.broadcastBreakEvent(hand));
-                    level.playSound(null, pos, SoundEvents.ANVIL_USE, SoundSource.BLOCKS, 1.0f, 0.8f + level.random.nextFloat() * 0.4f);
+                    level.playSound(null, pos, YTechSoundEvents.BRONZE_ANVIL_USE.get(), SoundSource.BLOCKS, 1.0f, 0.8f + level.random.nextFloat() * 0.4f);
                 }
             }